requirements.txt
Django==3.0.7
djangorestframework==3.11.0
Celery==4.4.6
Pillow==7.1.2
redis==3.5.3
requests==2.24.0
